requirements.txt
pytest
paramiko
pyyaml
numpy
gymnasium
prettytable
docutils
python-dotenv
openai
requests
